@@ -10,7 +10,7 @@ description: Arayüzü yapmaya çalışalım
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -29,6 +29,13 @@ dependencies:
   auto_size_text: 
   insta_like_button: ^0.1.1
   flutter_search_bar: ^2.1.0
+  json_annotation: any
+  json_serializable: any
+  dio:
+  http:
+  intl:
+  url_launcher: 
+  
 dev_dependencies:
   flutter_test:
     sdk: flutter
