@@ -1,8 +1,8 @@
 class ImageConstants {
-  static ImageConstants _instance;
+  static ImageConstants? _instance;
   static ImageConstants get instance {
     _instance ??= _instance = ImageConstants._init();
-    return _instance;
+    return _instance!;
   }
 
   ImageConstants._init();
