@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
 
 abstract class IAppTheme {
-  ThemeData theme;
+  ThemeData? theme;
 }
