@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_bounce/flutter_bounce.dart';
+import 'package:flutter_moda_app/design/constants/extension/context_extension.dart';
 import 'package:flutter_moda_app/design/constants/image/image_constants.dart';
 import 'package:flutter_svg/svg.dart';
 
@@ -8,8 +9,8 @@ class ProfilePage extends StatefulWidget {
   final Function onnext;
   static String routeName = 'CompleteProfileScreen';
   const ProfilePage({
-    Key key,
-    this.onnext,
+    Key? key,
+    required this.onnext,
   }) : super(key: key);
 
   @override
@@ -29,12 +30,15 @@ class _ProfilePageState extends State<ProfilePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.grey[50].withOpacity(0.9),
+      backgroundColor: Colors.grey[50]!.withOpacity(0.9),
       appBar: AppBar(
         backgroundColor: Colors.black.withOpacity(0.9),
         title: Text(
           "Profil",
-          style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
+          style: context.textTheme.headline6!.copyWith(
+            fontWeight: FontWeight.w500,
+            color: context.colors.secondary,
+          ),
         ),
         // ACTIONS KISMININ SADECE KAYITLI PROFİL HESABI OLANLARDA GÖRÜNECEK!!!
         // actions: <Widget>[
@@ -55,8 +59,15 @@ class _ProfilePageState extends State<ProfilePage> {
         //       Icons.exit_to_app,
         //       color: Colors.white,
         //     ),
-        //     onPressed: () async {},
-        //   )
+        //     onPressed: () {
+        //       showDialog(
+        //         context: context,
+        //         builder: (BuildContext context) {
+        //           return buildAlert(context);
+        //         },
+        //       );
+        //     },
+        //   ),
         // ],
         centerTitle: true,
         iconTheme: IconThemeData(
@@ -95,4 +106,33 @@ class _ProfilePageState extends State<ProfilePage> {
       ),
     );
   }
+  // ACTIONS KISMININ SADECE KAYITLI PROFİL HESABI OLANLARDA GÖRÜNECEK!!!
+  // Widget buildAlert(BuildContext context) {
+  //   return AlertDialog(
+  //     content: Text(
+  //       "Uygulamak çıkmak istediğine emin misin?",
+  //       style: TextStyle(color: Colors.black),
+  //     ),
+  //     actions: [
+  //       TextButton(
+  //         child: Text(
+  //           "Evet",
+  //           style: TextStyle(color: Colors.black),
+  //         ),
+  //         onPressed: () {
+  //           LocaleManager.instance.clearAll();
+  //           NavigationService.instance.navigateToPageClear(path: "/login");
+  //         },
+  //       ),
+  //       TextButton(
+  //           child: Text(
+  //             "Hayır",
+  //             style: TextStyle(color: Colors.red),
+  //           ),
+  //           onPressed: () {
+  //             Navigator.pop(context);
+  //           })
+  //     ],
+  //   );
+  // }
 }
