@@ -1,16 +1,14 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_moda_app/design/init/theme/app_theme_light.dart';
 import 'design/constants/navigation/navigation_constants.dart';
 import 'design/init/navigation/navigation_route.dart';
-import 'design/theme/app_theme_light.dart';
 
 void main() {
-  runApp(const Seepeaks());
+  runApp(Seepeaks());
 }
 
 class Seepeaks extends StatelessWidget {
-  const Seepeaks({Key key}) : super(key: key);
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
